@@ -1,73 +1,78 @@
 {
-  "name": "react-crud-ts",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "@material-ui/core": "^4.9.11",
-    "formik": "^2.1.4",
-    "localforage": "^1.7.3",
-    "lodash": "^4.17.15",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1",
-    "redux": "^4.0.5",
-    "redux-observable": "^1.2.0",
-    "redux-persist": "^6.0.0",
-    "rxjs": "^6.5.5",
-    "yup": "^0.28.3"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "@types/jest": "^24.9.1",
-    "@types/lodash": "^4.14.150",
-    "@types/material-ui": "^0.21.7",
-    "@types/node": "^12.12.36",
-    "@types/prop-types": "^15.7.3",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.6",
-    "@types/react-redux": "^7.1.7",
-    "@types/react-router-dom": "^5.1.4",
-    "@types/yup": "^0.26.37",
-    "@typescript-eslint/eslint-plugin": "^2.28.0",
-    "@typescript-eslint/parser": "^2.28.0",
-    "eslint": "^6.8.0",
-    "eslint-config-airbnb": "^18.1.0",
-    "eslint-config-prettier": "^6.10.1",
-    "eslint-import-resolver-typescript": "^2.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-json": "^2.1.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.19.0",
-    "prettier": "^2.0.4",
-    "redux-devtools-extension": "^2.13.8",
-    "typescript": "^3.8.3"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint ./src --ext .js,.jsx,.ts,.tsx"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  }
+	"name": "react-crud-ts",
+	"version": "0.1.0",
+	"private": true,
+	"dependencies": {
+		"@emotion/react": "^11.11.1",
+		"@emotion/styled": "^11.11.0",
+		"@mui/material": "^5.14.19",
+		"@mui/system": "^5.14.19",
+		"@typescript-eslint/eslint-plugin": "^6.13.1",
+		"@typescript-eslint/parser": "^6.13.1",
+		"formik": "^2.4.5",
+		"localforage": "^1.10.0",
+		"lodash": "^4.17.21",
+		"prop-types": "^15.8.1",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-redux": "^8.1.3",
+		"react-router-dom": "^6.20.1",
+		"react-scripts": "^5.0.1",
+		"redux": "^4.2.1",
+		"redux-observable": "^2.0.0",
+		"redux-persist": "^6.0.0",
+		"rxjs": "^7.8.1",
+		"yup": "^1.3.2"
+	},
+	"scripts": {
+		"start": "react-scripts start",
+		"build": "react-scripts build",
+		"test": "react-scripts test",
+		"eject": "react-scripts eject",
+		"lint": "eslint ./src --ext .js,.jsx,.ts,.tsx"
+	},
+	"eslintConfig": {
+		"extends": "react-app"
+	},
+	"browserslist": {
+		"production": [
+			">0.2%",
+			"not dead",
+			"not op_mini all"
+		],
+		"development": [
+			"last 1 chrome version",
+			"last 1 firefox version",
+			"last 1 safari version"
+		]
+	},
+	"Transform": "AWS::Serverless-2016-10-31",
+	"devDependencies": {
+		"@redux-devtools/extension": "^3.2.6",
+		"@testing-library/jest-dom": "^6.1.5",
+		"@testing-library/react": "^14.1.2",
+		"@testing-library/user-event": "^14.5.1",
+		"@types/jest": "^29.5.10",
+		"@types/lodash": "^4.14.202",
+		"@types/node": "^20.10.2",
+		"@types/prop-types": "^15.7.11",
+		"@types/react": "^18.2.40",
+		"@types/react-dom": "^18.2.17",
+		"@types/react-redux": "^7.1.31",
+		"@types/react-router-dom": "^5.3.3",
+		"@types/redux": "^3.6.0",
+		"@types/redux-persist": "^4.3.1",
+		"@types/yup": "^0.32.0",
+		"eslint": "^8.55.0",
+		"eslint-config-airbnb": "^19.0.4",
+		"eslint-config-prettier": "^9.0.0",
+		"eslint-import-resolver-typescript": "^3.6.1",
+		"eslint-plugin-import": "^2.29.0",
+		"eslint-plugin-json": "^3.1.0",
+		"eslint-plugin-jsx-a11y": "^6.8.0",
+		"eslint-plugin-prettier": "^5.0.1",
+		"eslint-plugin-react": "^7.33.2",
+		"jest": "^29.7.0",
+		"prettier": "^3.1.0"
+	}
 }
