@@ -1,6 +1,6 @@
 import React, { memo } from 'react';
-import Typography from '@material-ui/core/Typography';
-import Link from '@material-ui/core/Link';
+import Typography from '@mui/material/Typography';
+import Link from '@mui/material/Link';
 
 const Copyright = () => (
   <Typography variant="body2" color="textSecondary" align="center">
