@@ -1,10 +1,11 @@
 import { Action } from 'redux';
-import { ActionsObservable, ofType, StateObservable } from 'redux-observable';
+import { ofType, StateObservable } from 'redux-observable';
 import { mergeMap, map } from 'rxjs/operators';
 import { getJSON } from 'helper/request';
 import { userActionType, setUsers } from 'action/user';
+import { Observable } from 'rxjs';
 
-export const usersEpic = (action$: ActionsObservable<Action>, _store$: StateObservable<any>, _deps: any) =>
+export const usersEpic = (action$: Observable<Action<any>>, _store$: StateObservable<any>, _deps: any) =>
   action$.pipe(
     ofType(userActionType.pending),
     mergeMap(() =>
