@@ -1,8 +1,10 @@
 import React, { StrictMode } from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
+import { ThemeProvider } from '@emotion/react';
 import { PersistGate } from 'redux-persist/integration/react';
 import './index.css';
+import theme from './theme';
 import configureStore from './store';
 import Loading from 'component/loading';
 import App from './App';
@@ -13,9 +15,11 @@ import App from './App';
   ReactDOM.render(
     <StrictMode>
       <Provider store={store}>
-        <PersistGate loading={<Loading />} persistor={persistor}>
-          <App />
-        </PersistGate>
+        <ThemeProvider theme={theme}>
+          <PersistGate loading={<Loading />} persistor={persistor}>
+            <App />
+          </PersistGate>
+        </ThemeProvider>
       </Provider>
     </StrictMode>,
     document.getElementById('app')
